+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.2] Q1 2023
+ - report network status:
+   - switched to include filter
+   - fixed issue with empty filter
+   - report vectors of addresses
+
 ## [0.7.1] Q1 2023
  - added audit exception for RUSTSEC-2020-0071